broadcom: Add vc5 CLIF dumping
This will be usable with "VC5_DEBUG=clif" on the vc5 driver to stream a
CLIF file (the Broadcom equivalent of i965's AUB) to stderr.  I haven't
tested that this is actually usable with the internal CLIF-consuming
tools, but is close enough as a baseline and is useful for visually
inspecting the command stream.