nir: Move vc4's alpha test lowering to core NIR.
I've been doing this inside of vc4, but it seems like a pass that may be
useful for other drivers (Intel has a related path for pre-gen6 with MRT,
and freedreno had a TGSI path for it at one point).

This required defining a common enum for the standard comparison
functions, but other lowering passes are likely to also want that enum.